requirements.txt
# Core dependencies
playwright==1.51.0
python-dotenv==1.0.0
openai==1.69.0
aiohttp==3.8.5

# Additional utilities
2captcha-python==1.2.0